requirements_final.txt
black
python-dotenv
requests
flask
pytest
pytest-mock
numpy
pandas
matplotlib
werkzeug
sqlalchemy